requirements.txt
alembic
fastapi
uvicorn
sqlalchemy
pymysql
passlib==1.7.4
bcrypt==3.2.0
pyjwt
python-multipart
psycopg2-binary
redis